Tests: extract shared MinimalSoftwareWindow scaffolding
The TestPlatform / WINDOW thread-local / setup() block was copy-pasted
across three integration tests; centralising it makes future
software-renderer regression tests trivial to add.
